+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/signaling/message/MessageConverter.kt
@@ -16,7 +16,7 @@ class MessageConverter {
 
         @JvmOverloads
         fun buildConnectMessage(role: SoraChannelRole,
-                                channelId: String?,
+                                channelId: String,
                                 dataChannelSignaling: Boolean?,
                                 ignoreDisconnectWebSocket: Boolean?,
                                 mediaOption: SoraMediaOption,